+++ src/Tag/BbCommentaar.php
@@ -1,27 +1,28 @@
 <?php
 
-namespace CsrDelft\bb\tag;
+namespace CsrDelft\BbParser\Tag;
 
-use CsrDelft\bb\BbTag;
+use CsrDelft\BbParser\BbTag;
 
 /**
  * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
  * @since 27/03/2019
  */
-class BbUnderline extends BbTag
+class BbCommentaar extends BbTag
 {
+
     public static function getTagName(): string
     {
-        return 'u';
+        return 'commentaar';
     }
 
     public function parse($arguments = []): void
     {
-        $this->readContent();
+        $this->readContent([], false);
     }
 
     public function render(): string
     {
-        return '<ins class="onderstreept bb-tag-u">' . $this->getContent() . '</ins>';
+        return '';
     }
 }
