+++ src/Tag/BbLishort.php
@@ -1,22 +1,23 @@
 <?php
 
-namespace CsrDelft\bb\tag;
+namespace CsrDelft\BbParser\Tag;
 
-use CsrDelft\bb\BbTag;
+use CsrDelft\BbParser\BbTag;
 
 /**
- * List item
+ * List item (short)
  *
  * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
  * @since 27/03/2019
- * @example [li]Item[/li]
+ * @example [lishort]First item
+ * @example [*]Next item
  */
-class BbListItem extends BbTag
+class BbLishort extends BbTag
 {
 
-    public static function getTagName(): string
+    public static function getTagName(): array
     {
-        return 'li';
+        return ['lishort', '*'];
     }
 
     public function render(): string
@@ -36,6 +37,6 @@ class BbListItem extends BbTag
 
     public function parse($arguments = []): void
     {
-        $this->readContent();
+        $this->setChildren($this->parser->parseArray(['[br]']));
     }
 }
