+++ src/Tag/BbUnderline.php
@@ -1,28 +1,27 @@
 <?php
 
-namespace CsrDelft\bb\tag;
+namespace CsrDelft\BbParser\Tag;
 
-use CsrDelft\bb\BbTag;
+use CsrDelft\BbParser\BbTag;
 
 /**
  * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
  * @since 27/03/2019
  */
-class BbCommentaar extends BbTag
+class BbUnderline extends BbTag
 {
-
     public static function getTagName(): string
     {
-        return 'commentaar';
+        return 'u';
     }
 
     public function parse($arguments = []): void
     {
-        $this->readContent([], false);
+        $this->readContent();
     }
 
     public function render(): string
     {
-        return '';
+        return '<ins class="onderstreept bb-tag-u">' . $this->getContent() . '</ins>';
     }
 }
