+++ tests/PlainTagsTest.php
@@ -1,9 +1,9 @@
 <?php
 
 
-use CsrDelft\bb\BbEnv;
-use CsrDelft\bb\DefaultParser;
-use CsrDelft\bb\test\VarDriverPlatformIndependent;
+use CsrDelft\BbParser\BbEnv;
+use CsrDelft\BbParser\DefaultParser;
+use CsrDelft\BbParser\Test\VarDriverPlatformIndependent;
 use PHPUnit\Framework\TestCase;
 use Spatie\Snapshots\MatchesSnapshots;
